Change variable drvsel to LONG to properly support more than 16 drives.

+++ bdos/fs.h
@@ -348,7 +348,7 @@ DTAINFO
 extern  DND     *dirtbl[];
 extern  DMD     *drvtbl[];
 extern  char    diruse[];
-extern  int     drvsel;
+extern  LONG    drvsel;
 extern  PD      *run;
 extern  int     logmsk[];
 extern  FTAB    sft[];
